pasrseFloat!

+++ src/components/Text.js
@@ -9,7 +9,7 @@ export default function Text({ number, index, setCalc, showRes }) {
             setValue(event.target.value)
 
             setCalc(old => {
-            old[index] = parseInt(event.target.value)
+            old[index] = parseFloat(event.target.value)
             return old
             })
